core: Visual Studio compiler warnings (based on Robert Ammon's patch)

+++ src/core/examples/hash-example.cc
@@ -359,7 +359,7 @@ public:
     //
     // Number of buckets = k = 2^bits
     long double k32 = 0xFFFFFFFF;
-    long double k64 = 0xFFFFFFFFFFFFFFFFULL;
+    long double k64 = static_cast<long double> (0xFFFFFFFFFFFFFFFFULL);
 
     long double n = m_nphrases;
     long double Ec32 = n * (n - 1) / ( 2 * k32) * (1 - (n - 2)/(3 * k32));
