+++ src/core/model/nstime.h
@@ -394,7 +394,7 @@ public:
   }
   inline double GetDouble (void) const
   {
-    return m_data;
+    return static_cast<double> (m_data);
   }
   inline int64_t GetInteger (void) const
   {
